+++ server/src/parse.ts
@@ -31,9 +31,9 @@ export interface ComponentInfo {
 // Regex to match line components:
 // ^
 // (?<label>                           Label:
-//   ([^:\s;*=]+:?)                    - anything at start of line - optional colon
+//   ([^:\s;*=]+:?:?)                  - anything at start of line - optional colon
 //   |                                   or...
-//   (\s+[^:\s;*=]+:)                  - can have leading whitespace with colon present
+//   (\s+[^:\s;*=]+::?)                - can have leading whitespace with colon present
 // )?
 // (\s*                                Instruction or directive:
 //   (
@@ -59,7 +59,7 @@ export interface ComponentInfo {
 // (\s*(?<comment>.+))?                Comment (any trailing text)
 // $
 const pattern =
-  /^(?<label>([^:\s;*=]+:?)|(\s+[^:\s;*=]+:))?(\s*(((?<mnemonic1>\.?(nop|reset|rte|rtr|rts|trapv|illegal|clrfo|clrso|comment|einline|even|inline|list|mexit|nolist|nopage|odd|page|popsection|pushsection|rsreset|endif|endc|else|elseif|endm|endr|erem))(?<size1>\.[a-z0-9_.]*)?)|((?<mnemonic>([^\s.,;*=]+|=))(?<size>\.[^\s.,;*]*)?(\s*(?<operands>([^\s;"',]+|"([^"]*)"?|'([^']*)'?)(,\s*[^\s;"',]*|"([^"]*)"?|'([^']*)'?)*))?)))?(\s*(?<comment>.+))?$/i;
+  /^(?<label>([^:\s;*=]+:?:?)|(\s+[^:\s;*=]+::?))?(\s*(((?<mnemonic1>\.?(nop|reset|rte|rtr|rts|trapv|illegal|clrfo|clrso|comment|einline|even|inline|list|mexit|nolist|nopage|odd|page|popsection|pushsection|rsreset|endif|endc|else|elseif|endm|endr|erem))(?<size1>\.[a-z0-9_.]*)?)|((?<mnemonic>([^\s.,;*=]+|=))(?<size>\.[^\s.,;*]*)?(\s*(?<operands>([^\s;"',]+|"([^"]*)"?|'([^']*)'?)(,\s*[^\s;"',]*|"([^"]*)"?|'([^']*)'?)*))?)))?(\s*(?<comment>.+))?$/i;
 
 /**
  * Parse a single line of source code into positional components
@@ -75,7 +75,7 @@ export function parseLine(text: string): ParsedLine {
 
     if (groups.label) {
       let value = groups.label.trim();
-      if (value.endsWith(":")) {
+      while (value.endsWith(":")) {
         value = value.substring(0, value.length - 1);
       }
       const start = text.indexOf(value);